+++ Thunderstore/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### v1.0.1
+
+- Add autoReorderOnLeftMoon config, so you don't have to type command each day
+- Add default reorder type config
+
 ### v1.0.0
 
 - initial release